refactor(preview): extract cancellable process helpers
