+++ src/preview/mod.rs
@@ -8,6 +8,7 @@ mod directory;
 mod dispatch;
 mod document;
 mod markdown;
+mod process;
 mod structured;
 mod text;
 mod types;
